+++ src/Handlers/MiraiApiHandler.php
@@ -21,6 +21,9 @@ use Swlib\Saber;
  */
 class MiraiApiHandler
 {
+    /** @var Config DiceRobot config. */
+    protected Config $config;
+
     /** @var Saber Client pool. */
     protected Saber $pool;
 
@@ -30,11 +33,14 @@ class MiraiApiHandler
     /**
      * The constructor.
      *
+     * @param Config $config DiceRobot config.
      * @param LoggerFactory $loggerFactory Logger factory.
      */
-    public function __construct(LoggerFactory $loggerFactory)
+    public function __construct(Config $config, LoggerFactory $loggerFactory)
     {
-        $this->logger = $loggerFactory->create("Handler");
+        $this->config = $config;
+
+        $this->logger = $loggerFactory->create("MiraiAPI");
 
         $this->logger->debug("Mirai API handler created.");
     }
@@ -49,18 +55,16 @@ class MiraiApiHandler
 
     /**
      * Initialize handler.
-     *
-     * @param Config $config DiceRobot config.
      */
-    public function initialize(Config $config): void
+    public function initialize(): void
     {
         $this->pool = Saber::create([
             "base_uri" =>
-                "http://{$config->getString("mirai.server.host")}:{$config->getString("mirai.server.port")}",
+                "http://{$this->config->getString("mirai.server.host")}:{$this->config->getString("mirai.server.port")}",
             "use_pool" => true,
             "headers" => [
                 "Content-Type" => ContentType::JSON,
-                "User-Agent" => "DiceRobot/{$config->getString("dicerobot.version")}"
+                "User-Agent" => "DiceRobot/{$this->config->getString("dicerobot.version")}"
             ],
             "before" => function (Saber\Request $request) {
                 $this->logger->debug("Send to {$request->getUri()}, content: {$request->getBody()}");