+++ src/Handlers/PanelHandler.php
@@ -6,14 +6,16 @@ namespace DiceRobot\Handlers;
 
 use DiceRobot\App;
 use DiceRobot\Data\Config;
+use DiceRobot\Data\CustomConfig;
 use DiceRobot\Factory\{LoggerFactory, ResponseFactory};
-use DiceRobot\Server;
-use DiceRobot\Service\{RobotService, StatisticsService};
+use DiceRobot\Service\{LogService, ResourceService, RobotService, StatisticsService};
 use DiceRobot\Util\Updater;
 use Psr\Container\ContainerInterface;
 use Psr\Log\LoggerInterface;
 use Swoole\Coroutine\System;
 use Swoole\Http\Response;
+use Swoole\Process;
+use Throwable;
 
 /**
  * Class PanelHandler
@@ -33,8 +35,11 @@ class PanelHandler
     /** @var App Application. */
     protected App $app;
 
-    /** @var Server DiceRobot HTTP server. */
-    protected Server $server;
+    /** @var LogService Log service. */
+    protected LogService $log;
+
+    /** @var ResourceService Resource service. */
+    protected ResourceService $resource;
 
     /** @var RobotService Robot service. */
     protected RobotService $robot;
@@ -54,6 +59,8 @@ class PanelHandler
      * @param ContainerInterface $container Container.
      * @param Config $config DiceRobot config.
      * @param App $app Application.
+     * @param LogService $log Log service.
+     * @param ResourceService $resource Resource service.
      * @param RobotService $robot Robot service.
      * @param StatisticsService $statistics Statistics service.
      * @param ResponseFactory $responseFactory HTTP response factory.
@@ -63,6 +70,8 @@ class PanelHandler
         ContainerInterface $container,
         Config $config,
         App $app,
+        LogService $log,
+        ResourceService $resource,
         RobotService $robot,
         StatisticsService $statistics,
         ResponseFactory $responseFactory,
@@ -71,11 +80,13 @@ class PanelHandler
         $this->container = $container;
         $this->config = $config;
         $this->app = $app;
+        $this->log = $log;
+        $this->resource = $resource;
         $this->robot = $robot;
         $this->statistics = $statistics;
         $this->responseFactory = $responseFactory;
 
-        $this->logger = $loggerFactory->create("Handler");
+        $this->logger = $loggerFactory->create("Panel");
 
         $this->logger->debug("Panel handler created.");
     }
@@ -89,15 +100,121 @@ class PanelHandler
     }
 
     /**
-     * Initialize handler.
+     * Handle panel request.
      *
-     * @param Server $server DiceRobot HTTP server.
+     * @param string $method Request method.
+     * @param string $uri Request URL.
+     * @param string[] $queryParams Query parameters.
+     * @param string $content Request content.
+     * @param string $contentType Request content type.
+     * @param Response $response Swoole response.
+     *
+     * @return Response Swoole response.
      */
-    public function initialize(Server $server): void
-    {
-        $this->server = $server;
+    public function handle(
+        string $method,
+        string $uri,
+        array $queryParams,
+        string $content,
+        string $contentType,
+        Response $response
+    ): Response {
+        try {
+            $this->route($method, $uri, $queryParams, $content, $contentType, $response);
+        } catch (Throwable $t) {
+            $details = sprintf(
+                "Type: %s\nCode: %s\nMessage: %s\nFile: %s\nLine: %s\nTrace: %s",
+                get_class($t),
+                $t->getCode(),
+                $t->getMessage(),
+                $t->getFile(),
+                $t->getLine(),
+                $t->getTraceAsString()
+            );
 
-        $this->logger->info("Panel handler initialized.");
+            $this->logger->error("Panel request failed, unexpected exception occurred:\n{$details}.");
+
+            $this->error($response);
+        }
+
+        return $response;
+    }
+
+    /**
+     * Route request.
+     *
+     * @param string $method Request method.
+     * @param string $uri Request URL.
+     * @param string[] $queryParams Query parameters.
+     * @param string $content Request content.
+     * @param string $contentType Request content type.
+     * @param Response $response Swoole response.
+     *
+     * @return Response Swoole response.
+     */
+    protected function route(
+        string $method,
+        string $uri,
+        array $queryParams,
+        string $content,
+        string $contentType,
+        Response $response
+    ): Response {
+        if ($method == "OPTIONS") {
+            $this->preflight($response);
+        } elseif ($method == "POST" && $contentType == "application/json") {
+            if ($uri == "/config") {
+                $this->setConfig($content, $response);
+            } elseif ($uri == "/mirai/update") {
+                $this->updateMirai($content, $response);
+            } else {
+                $this->notFound($response);
+            }
+        } elseif ($method == "GET") {
+            if ($uri == "/connect") {
+                $this->connect($response);
+            } elseif ($uri == "/pause") {
+                $this->pause($response);
+            } elseif ($uri == "/run") {
+                $this->rerun($response);
+            } elseif ($uri == "/reload") {
+                $this->reload($response);
+            } elseif ($uri == "/stop") {
+                $this->stop($response);
+            } elseif ($uri == "/restart") {
+                $this->restart($response);
+            } elseif ($uri == "/update") {
+                $this->update($response);
+            } elseif ($uri == "/profile") {
+                $this->getProfile($response);
+            } elseif ($uri == "/status") {
+                $this->getStatus($response);
+            } elseif ($uri == "/statistics") {
+                $this->getStatistics($response);
+            } elseif ($uri == "/config") {
+                $this->getConfig($response);
+            } elseif ($uri == "/logs") {
+                $this->getLogList($response);
+            } elseif ($uri == "/log") {
+                $this->getLog($queryParams, $response);
+            } elseif ($uri == "/skeleton/update") {
+                $this->updateSkeleton($response);
+            } elseif ($uri == "/mirai/status") {
+                $this->getMiraiStatus($response);
+            } elseif ($uri == "/mirai/start") {
+                $this->startMirai($response);
+            } elseif ($uri == "/mirai/stop") {
+                $this->stopMirai($response);
+            } elseif ($uri == "/mirai/restart") {
+                $this->restartMirai($response);
+            } else {
+                $this->notFound($response);
+            }
+        } else {
+            $this->notFound($response);
+        }
+
+        return $response;
     }
 
     /**
@@ -107,7 +224,7 @@ class PanelHandler
      *
      * @return Response HTTP response.
      */
-    public function preflight(Response $response): Response
+    protected function preflight(Response $response): Response
     {
         return $this->responseFactory->createPreflight($response);
     }
@@ -119,126 +236,21 @@ class PanelHandler
      *
      * @return Response HTTP response.
      */
-    public function notFound(Response $response): Response
+    protected function notFound(Response $response): Response
     {
         return $this->responseFactory->createNotFound($response);
     }
 
     /**
-     * Get profile.
+     * 500 Internal Server Error.
      *
      * @param Response $response HTTP response.
      *
      * @return Response HTTP response.
      */
-    public function getProfile(Response $response): Response
+    protected function error(Response $response): Response
     {
-        $this->logger->info("HTTP request received, get profile.");
-
-        $data = [
-            "id" => $this->robot->getId(),
-            "nickname" => $this->robot->getNickname(),
-            "friends" => $this->robot->getFriendCount(),
-            "groups" => $this->robot->getGroupCount(),
-            "startup" => DICEROBOT_STARTUP,
-            "version" => [
-                "dicerobot" => DICEROBOT_VERSION,
-                "mirai" => $this->robot->getVersion()
-            ]
-        ];
-
-        return $this->responseFactory->create(0, $data, $response);
-    }
-
-    /**
-     * Get status.
-     *
-     * @param Response $response HTTP response.
-     *
-     * @return Response HTTP response.
-     */
-    public function getStatus(Response $response): Response
-    {
-        $this->logger->info("HTTP request received, get application status.");
-
-        $appStatus = $this->app->getStatus()->getValue();
-        $code = -1;
-
-        extract(System::exec(
-            "/bin/systemctl status {$this->config->getString("dicerobot.service.name")}"
-        ), EXTR_OVERWRITE);
-
-        if ($code == 4) {
-            $serviceStatus = -2;
-        } elseif ($code == 3) {
-            $serviceStatus = -1;
-        } elseif ($code == 0) {
-            $serviceStatus = 0;
-        } else {
-            $serviceStatus = -3;
-        }
-
-        return $this->responseFactory->create(0, ["app" => $appStatus, "service" => $serviceStatus], $response);
-    }
-
-    /**
-     * Get statistics.
-     *
-     * @param Response $response HTTP response.
-     *
-     * @return Response HTTP response.
-     */
-    public function getStatistics(Response $response): Response
-    {
-        $this->logger->info("HTTP request received, get statistics.");
-
-        $data = $this->statistics->getSortedData();
-
-        return $this->responseFactory->create(0, $data, $response);
-    }
-
-    /**
-     * Get DiceRobot config.
-     *
-     * @param Response $response HTTP response.
-     *
-     * @return Response HTTP response.
-     */
-    public function getConfig(Response $response): Response
-    {
-        $this->logger->info("HTTP request received, get config.");
-
-        $data = [
-            "strategy" => $this->config->getArray("strategy"),
-            "order" => $this->config->getArray("order"),
-            "reply" => $this->config->getArray("reply"),
-            "errMsg" => $this->config->getArray("errMsg"),
-        ];
-
-        return $this->responseFactory->create(0, $data, $response);
-    }
-
-    /**
-     * Set panel config.
-     *
-     * @param string $content HTTP request content.
-     * @param Response $response HTTP response.
-     *
-     * @return Response HTTP response.
-     */
-    public function setConfig(string $content, Response $response): Response
-    {
-        $this->logger->info("HTTP request received, set config.");
-
-        if (!is_array($data = json_decode($content, true))) {
-            return $this->responseFactory->create(-1060, null, $response);
-        }
-
-        if (!$this->app->setConfig($data)) {
-            return $this->responseFactory->create(-1061, null, $response);
-        }
-
-        return $this->responseFactory->create(0, null, $response);
+        return $this->responseFactory->createError($response);
     }
 
     /**
@@ -248,7 +260,7 @@ class PanelHandler
      *
      * @return Response HTTP response.
      */
-    public function connect(Response $response): Response
+    protected function connect(Response $response): Response
     {
         $this->logger->info("HTTP request received, connect to application.");
 
@@ -262,7 +274,7 @@ class PanelHandler
      *
      * @return Response HTTP response.
      */
-    public function pause(Response $response): Response
+    protected function pause(Response $response): Response
     {
         $this->logger->notice("HTTP request received, pause application.");
 
@@ -284,7 +296,7 @@ class PanelHandler
      *
      * @return Response HTTP response.
      */
-    public function rerun(Response $response): Response
+    protected function rerun(Response $response): Response
     {
         $this->logger->notice("HTTP request received, rerun application.");
 
@@ -308,7 +320,7 @@ class PanelHandler
      *
      * @return Response HTTP response.
      */
-    public function reload(Response $response): Response
+    protected function reload(Response $response): Response
     {
         $this->logger->notice("HTTP request received, reload application.");
 
@@ -326,7 +338,7 @@ class PanelHandler
      *
      * @param Response $response HTTP response.
      */
-    public function stop(Response $response): void
+    protected function stop(Response $response): void
     {
         $this->logger->notice("HTTP request received, stop application.");
 
@@ -338,7 +350,7 @@ class PanelHandler
 
         $this->responseFactory->create($code, null, $response)->end();
 
-        $this->server->stop();
+        Process::kill(getmypid(), 10);  // Send SIGUSR1
     }
 
     /**
@@ -346,7 +358,7 @@ class PanelHandler
      *
      * @param Response $response HTTP response.
      */
-    public function restart(Response $response): void
+    protected function restart(Response $response): void
     {
         $this->logger->notice("HTTP request received, restart application.");
 
@@ -372,7 +384,7 @@ class PanelHandler
      *
      * @return Response HTTP response.
      */
-    public function update(Response $response): Response
+    protected function update(Response $response): Response
     {
         $this->logger->notice("HTTP request received, update DiceRobot.");
 
@@ -407,6 +419,162 @@ class PanelHandler
     }
 
     /**
+     * Get profile.
+     *
+     * @param Response $response HTTP response.
+     *
+     * @return Response HTTP response.
+     */
+    protected function getProfile(Response $response): Response
+    {
+        $this->logger->info("HTTP request received, get profile.");
+
+        $data = [
+            "id" => $this->robot->getId(),
+            "nickname" => $this->robot->getNickname(),
+            "friends" => $this->robot->getFriendCount(),
+            "groups" => $this->robot->getGroupCount(),
+            "startup" => DICEROBOT_STARTUP,
+            "version" => [
+                "dicerobot" => DICEROBOT_VERSION,
+                "mirai" => $this->robot->getVersion()
+            ]
+        ];
+
+        return $this->responseFactory->create(0, $data, $response);
+    }
+
+    /**
+     * Get status.
+     *
+     * @param Response $response HTTP response.
+     *
+     * @return Response HTTP response.
+     */
+    protected function getStatus(Response $response): Response
+    {
+        $this->logger->info("HTTP request received, get application status.");
+
+        $appStatus = $this->app->getStatus()->getValue();
+        $code = -1;
+
+        extract(System::exec(
+            "/bin/systemctl status {$this->config->getString("dicerobot.service.name")}"
+        ), EXTR_OVERWRITE);
+
+        if ($code == 4) {
+            $serviceStatus = -2;
+        } elseif ($code == 3) {
+            $serviceStatus = -1;
+        } elseif ($code == 0) {
+            $serviceStatus = 0;
+        } else {
+            $serviceStatus = -3;
+        }
+
+        return $this->responseFactory->create(0, ["app" => $appStatus, "service" => $serviceStatus], $response);
+    }
+
+    /**
+     * Get statistics.
+     *
+     * @param Response $response HTTP response.
+     *
+     * @return Response HTTP response.
+     */
+    protected function getStatistics(Response $response): Response
+    {
+        $this->logger->info("HTTP request received, get statistics.");
+
+        $data = $this->statistics->getSortedData();
+
+        return $this->responseFactory->create(0, $data, $response);
+    }
+
+    /**
+     * Get DiceRobot config.
+     *
+     * @param Response $response HTTP response.
+     *
+     * @return Response HTTP response.
+     */
+    protected function getConfig(Response $response): Response
+    {
+        $this->logger->info("HTTP request received, get config.");
+
+        $data = [
+            "strategy" => $this->config->getArray("strategy"),
+            "order" => $this->config->getArray("order"),
+            "reply" => $this->config->getArray("reply"),
+            "errMsg" => $this->config->getArray("errMsg"),
+        ];
+
+        return $this->responseFactory->create(0, $data, $response);
+    }
+
+    /**
+     * Set panel config.
+     *
+     * @param string $content HTTP request content.
+     * @param Response $response HTTP response.
+     *
+     * @return Response HTTP response.
+     *
+     * @noinspection PhpDocMissingThrowsInspection
+     */
+    protected function setConfig(string $content, Response $response): Response
+    {
+        $this->logger->info("HTTP request received, set config.");
+
+        if (!is_array($data = json_decode($content, true))) {
+            return $this->responseFactory->create(-1060, null, $response);
+        }
+
+        if (!$this->resource->getConfig()->setConfig($data)) {
+            return $this->responseFactory->create(-1061, null, $response);
+        }
+
+        $this->config->load($this->container->make(CustomConfig::class), $this->resource->getConfig());
+
+        return $this->responseFactory->create(0, null, $response);
+    }
+
+    /**
+     * Get DiceRobot log file list.
+     *
+     * @param Response $response HTTP response.
+     *
+     * @return Response HTTP response.
+     */
+    protected function getLogList(Response $response): Response
+    {
+        $this->logger->info("HTTP request received, get log list.");
+
+        $logs = $this->log->getLogList();
+
+        return $this->responseFactory->create(0, ["logs" => $logs], $response);
+    }
+
+    /**
+     * Get parsed DiceRobot log file content.
+     *
+     * @param array $params Query parameters.
+     * @param Response $response HTTP response.
+     *
+     * @return Response HTTP response.
+     */
+    protected function getLog(array $params, Response $response): Response
+    {
+        $this->logger->info("HTTP request received, get log.");
+
+        if (false === $log = $this->log->getLog($params["name"] ?? "")) {
+            return $this->responseFactory->create(-1080, null, $response);
+        }
+
+        return $this->responseFactory->create(0, ["log" => $log], $response);
+    }
+
+    /**
      * Update DiceRobot skeleton.
      *
      * @param Response $response HTTP response.
@@ -416,7 +584,7 @@ class PanelHandler
      * @noinspection PhpDocMissingThrowsInspection
      * @noinspection PhpUnhandledExceptionInspection
      */
-    public function updateSkeleton(Response $response): Response
+    protected function updateSkeleton(Response $response): Response
     {
         $this->logger->notice("HTTP request received, update skeleton.");
 
@@ -441,7 +609,7 @@ class PanelHandler
      *
      * @return Response HTTP response.
      */
-    public function getMiraiStatus(Response $response): Response
+    protected function getMiraiStatus(Response $response): Response
     {
         $this->logger->info("HTTP request received, get Mirai status.");
 
@@ -471,7 +639,7 @@ class PanelHandler
      *
      * @return Response HTTP response.
      */
-    public function startMirai(Response $response): Response
+    protected function startMirai(Response $response): Response
     {
         $this->logger->notice("HTTP request received, start Mirai.");
 
@@ -508,7 +676,7 @@ class PanelHandler
      *
      * @return Response HTTP response.
      */
-    public function stopMirai(Response $response): Response
+    protected function stopMirai(Response $response): Response
     {
         $this->logger->notice("HTTP request received, stop Mirai.");
 
@@ -545,7 +713,7 @@ class PanelHandler
      *
      * @return Response HTTP response.
      */
-    public function restartMirai(Response $response): Response
+    protected function restartMirai(Response $response): Response
     {
         $this->logger->notice("HTTP request received, restart Mirai.");
 
@@ -583,7 +751,7 @@ class PanelHandler
      *
      * @return Response HTTP response.
      */
-    public function updateMirai(string $content, Response $response): Response
+    protected function updateMirai(string $content, Response $response): Response
     {
         $this->logger->notice("HTTP request received, update Mirai.");